requirements.txt
boto3>=1.28.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
numpy>=1.24.0
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cpu
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
